feat(auth): add reusable LoginSocial component for Google and GitHub authentication
- Created a new LoginSocial component integrating Google and GitHub OAuth login flows.
- Implemented unified handler for social login redirection.
- Styled buttons with Tailwind for consistent UI design.
- Made the component reusable across both SignIn and SignUp pages.
