+++ frontend/src/pages/auth/SignIn.jsx
@@ -12,6 +12,7 @@ import BackToTopButton from '@/components/utils/BackToTopButton';
 import { motion } from 'framer-motion';
 import Channels from '@/components/channels/Channels';
 import Spinner from '@/components/spinner/Spinner';
+import LoginSocial from '@/components/utils/LoginSocial';
 
 export default function SignIn() {
   const navigate = useNavigate()
@@ -30,19 +31,6 @@ export default function SignIn() {
     handleSubmit,
     formState: { errors },
   } = useForm({ resolver: yupResolver(loginSchema) })
-
-  // Social login (Google & GitHub unified)
-  const handleSocialLogin = (provider) => {
-    const backendBaseUrl =
-      import.meta.env.VITE_URL_BACKEND || 'http://localhost:5001';
-
-    if (provider === 'google') {
-      window.location.href = `${backendBaseUrl}/auth/google`;
-    } else if (provider === 'github') {
-      // GitHub OAuth
-      window.location.href = `${backendBaseUrl}/auth/github`;
-    }
-  };
 
   // Traditional login
   const handleLoginSubmit = async (data) => {
@@ -212,33 +200,8 @@ export default function SignIn() {
                   )}
                 </button>
 
-                {/* Divider */}
-                <div className='flex items-center my-8'>
-                  <div className='flex-grow border-t border-gray-400'></div>
-                  <span className='mx-4 text-white text-sm'>
-                    Or sign in through
-                  </span>
-                  <div className='flex-grow border-t border-gray-400'></div>
-                </div>
-
                 {/* Social Buttons */}
-                <div className='flex justify-center gap-4 mb-9'>
-                  <button
-                    type='button'
-                    className='text-white bg-purple-600 hover:bg-purple-800 p-2 rounded-full cursor-pointer'
-                    onClick={() => handleSocialLogin('google')}
-                  >
-                    <FaGoogle className='w-5 h-5' />
-                  </button>
-
-                  <button
-                    type='button'
-                    className='text-white bg-purple-600 hover:bg-purple-800 p-2 rounded-full cursor-pointer'
-                    onClick={() => handleSocialLogin('github')}
-                  >
-                    <FaGithub className='w-5 h-5' />
-                  </button>
-                </div>
+                <LoginSocial />
 
                 {/* Sign Up Link */}
                 <div className='mt-6 flex flex-col items-center'>
